+++ tests/tests_client.py
@@ -15,7 +15,7 @@ from tests.tests_highscores import FILE_HIGHSCORES_FULL
 from tests.tests_house import FILE_HOUSE_FULL, FILE_HOUSE_LIST
 from tests.tests_kill_statistics import FILE_KILL_STATISTICS_FULL
 from tests.tests_leaderboards import FILE_LEADERBOARD_CURRENT
-from tests.tests_news import FILE_NEWS_LIST, FILE_NEWS_ARTICLE
+from tests.tests_news import FILE_NEWS_ARCHIVE_RESULTS, FILE_NEWS_ARTICLE
 from tests.tests_tibiapy import TestCommons
 from tests.tests_world import FILE_WORLD_FULL, FILE_WORLD_LIST
 from tibiapy import Client, Forbidden, NetworkError, HouseType
@@ -152,7 +152,7 @@ class TestClient(unittest.TestCase, TestCommons):
     @aioresponses()
     async def test_client_fetch_recent_news(self, mock):
         """Testing fetching recent nows"""
-        content = self.load_resource(FILE_NEWS_LIST)
+        content = self.load_resource(FILE_NEWS_ARCHIVE_RESULTS)
         mock.post(NewsArchive.get_url(), status=200, body=content)
         recent_news = await self.client.fetch_recent_news(30)