+++ tibiapy/utils.py
@@ -182,8 +182,9 @@ def parse_form_data_new(form: bs4.Tag):
         form_data.values[name] = selected_option.attrs.get("value") if selected_option else None
     for checkbox in form.select("input[type=checkbox]"):
         name = checkbox.attrs.get("name")
+        label = checkbox.parent.text
         value = checkbox.attrs.get("value")
-        form_data.available_options[name].append(value)
+        form_data.available_options[name][label] = value
         if checkbox.has_attr("checked"):
             form_data.values_multiple[name].append(value)
     for radio in form.select("input[type=radio]"):